backend/requirements.txt
Flask==3.1.1
python-dotenv 
psycopg2
playwright
requests
schedule
